# Add memoization by instance