--- a/src/test/testIntegrations.js
+++ b/src/test/testIntegrations.js
@@ -913,6 +913,6 @@ describe('Many-to-many relationship performance', () => {
         const end = new Date().getTime();
         const tookSeconds = (end - start) / 1000;
         console.log(`Removing ${removeCount} relations took ${tookSeconds}s`);
-        expect(tookSeconds).toBeLessThanOrEqual(process.env.TRAVIS ? 11.5 : 4);
+        expect(tookSeconds).toBeLessThanOrEqual(process.env.TRAVIS ? 12 : 4);
     });
 });