@B-Reif 2018-02-01
Just pushed a commit with a few of the fixes from above.

Out of curiosity about the performance impact, I ran `npm run test` a few times at this commit, and the last commit before any changes on the branch. The `testIntegrations` test has a few benchmarks and a total time.

Before changes: 11.912s, 10.555s, and 10.362s (avg 10.94s)
With changes: 11.359s, 10.429s, and 11.088s (avg 10.95s)

It seems like the difference in overall performance is negligible, at least as far as these tests are concerned.