@tiii 2018-02-03
Over application lifetime, we keep adding up models to our session.accessedModels object and therefore incrementally checking models that did not change for changes over time.

I was thinking if, for redux, we could flush the session.accessedModels after each application render, since our potential changes are then rendered to the dom and only changes to models that occured after the last render need to be checked for equality again.

By providing an API to reset session.accessedModels and a higher order component that's placed early in the react component tree that calls that reset function in componentDidUpdate() we could keep checking the number of accessedModels that need to be checked for equality to a minimum.

What are your thoughts on this? Am I missing a situation, where this may lead up to missing changes? Otherwise this sure would be a big performance improvement over app lifetime.