@tiii 2018-02-04
Turns out https://github.com/tommikaikkonen/redux-orm/pull/185#issuecomment-362857766 might be [relevant after all](https://github.com/tommikaikkonen/redux-orm/pull/185/files/2aec1fd523648341d92c063904cd76b0b63684d7#r165829431)...