@fatfatson 2018-02-09
a simple question:
how could we know the model doesn't change?
now I just make a deep-compare in shouldComponentUpdate, if we just move the check from there to  internal of redux-orm, it may be more convenient to use but the performance doesn't improve