@B-Reif 2018-02-12
Currently the comparison happens in the memoize function within redux-orm. Right now, this function just checks if the previous state's table is `===` equal to the next state's table. Per @tiii's changes, we'd like to do the same thing--a quick reference check--against each model _instance_.

Reference checking each instance is substantially more performant than deep-comparing all of the properties for each instance (as shouldComponentUpdate would be doing), along with the other benefits of memoization like caching computed data.