@B-Reif 2018-02-21
I made some changes to implement the `memoize` function per @tiii's ideas. On this last commit, the selectors are currently broken and don't re-run when new models are added to the table. For different tables, the same query can return a different set of instances. AFAIK, it's not possible to determine which instances a query will yield without actually running it.

For example, some selector that looks for `Book.all()` won't re-run when a new Book has been added. The first time, the selector stores which instances it got. The second time, it compares those instances and sees they're all equal, so it doesn't re-run the same query which would, if ran, yield a different set of instances.

We could just add a check like `lastModelState.items === nextModelState.items`. This would cause other cases to re-run the selector unnecessarily. A query that looks for `Book.withId(0)` would re-run whenever a new Book is added or removed.

I also feel like we should add a `createSelectorCreator` to allow developers to make the call about when to re-run if they'd like to.