@tiii 2018-02-21
I just refactored the existing code along with @Aryk's proposal, so all props go to him 👍 

I feel like the model.items equality check is the way to go. If the selector only uses withId it should run pretty fast, and in the end - since nothing changed - the references are still the same and react shouldn't rerender.

I guess we could also store the accessedModels per selector function and check if only withId is called, but that seems utterly complex and pretty much overkill.