@tiii 2018-02-21
I just realized we already memoize per selector. So we could also track the used query methods to find out if only withId is used.

I hope I find time to provide some test cases / documentation soon..