@B-Reif 2018-02-21
This implementation is still broken for ‘filter’ and ‘exclude’ queries, since they might hit new instances on each change.