@B-Reif 2018-02-21
To be honest, I'm not sure if there's a good strategy to apply across all possible query types. Some query types, like `filter`, have to actually examine every instance in the table. Others, like `withId`, only care about one given instance. Others still, like `first` or `at(index)`, are concerned with the ordering of the QuerySet.

I think what I'd like to do is:
- Expose a version of `createSelectorCreator` to allow developers to specify their own memoization function.
- Keep the changes that mark individual instances as accessed, so that users of `createSelectorCreator` can memoize to the instance level where it's appropriate to do so.
- Remove the changes to `memoize`, and keep the default memoization at the table level.

These changes would allow developers to take responsibility of memoization, and give developers the information they need to make meaningful memoization functions. But it would also leave the weak table-level default for most developers. And storing the instance ids in the map leaves queries (slightly) slower for everyone.

These changes would also retain most backward compatibility, excepting `session.accessedModels`, which will be a map instead of an array. I don't think `session.accessedModels` is intended as public API anyway.

The alternative would be to come up with a default memoization strategy that attempts to cover all possible use cases. This strategy would have to track which queries were used and their arguments to divine what level of the state it should be comparing at.