@Aryk 2018-02-22
@B-Reif when you said "This implementation is still broken for ‘filter’ and ‘exclude’ queries, since they might hit new instances on each change."

Are you saying that potentially new records could be added to the tables and hence the memoized result would not pick up the new records? Wasn't that handled before by just adding the model to accessed Models? Why don't we keep the old functionality in this way alongside the new per-id model. Maybe we could add in the map something like `{[modelName]: true}` or something like that.

When a filter or exclude happens you could add a flag to check the IDs of the table. You could use an efficient "===" check by looking at the state directly. It would be `db.Model.items` in the state which is an array of the IDs. If an add or remove happened that array would be invalidated.