@B-Reif 2018-02-22
That’s not the problem; the problem is that existing instances could change such that they now pass the filter, when they previously did not.

For example, suppose we have a filter query that looks for Book models with a name starting with “P”. The first run, it caches the applicable model instances. Then I add a “P” to some other instance. Our filter query then doesn’t re-run, because the change was outside the cached instances.