@Aryk 2018-02-22
Right, that is why I am proposing anytime you hit something that relies on "bypassed" or "skipped" records, that you put {Book: true} into the accessedModels. That is basically like old functionality.

The difference is that now, because we have instance level caching, when you render an individual book with withId on a connected component, you won't actually have to rerender that entire component, whereas even if you have a connected component on one instance with Book ID "5", it will rerender when Book Id "6" gets updated.